[PLAT-14366] Basic local provider test for master auto failover
Summary: Local provider test for auto-master failover.

Test Plan:
Ran the failover in a docker container.

```
[info] Test run com.yugabyte.yw.commissioner.tasks.local.AutoMasterFailoverLocalTest finished: 0 failed, 0 ignored, 1 total, 364.275s
[info] Passed: Total 1, Failed 0, Errors 0, Passed 1
```

Reviewers: sanketh, cwang, yshchetinin

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D36726

+++ managed/src/main/java/com/yugabyte/yw/common/LocalNodeManager.java
@@ -33,6 +33,7 @@ import com.yugabyte.yw.common.gflags.GFlagsUtil;
 import com.yugabyte.yw.common.gflags.SpecificGFlags;
 import com.yugabyte.yw.common.utils.FileUtils;
 import com.yugabyte.yw.forms.UniverseDefinitionTaskParams;
+import com.yugabyte.yw.forms.UniverseDefinitionTaskParams.UserIntent;
 import com.yugabyte.yw.models.Provider;
 import com.yugabyte.yw.models.ProviderDetails;
 import com.yugabyte.yw.models.Universe;
@@ -224,12 +225,33 @@ public class LocalNodeManager {
             });
   }
 
+  // This does not clear the process map.
   public void killProcess(String nodeName, UniverseTaskBase.ServerType serverType)
       throws IOException, InterruptedException {
     NodeInfo nodeInfo = nodesByNameMap.get(nodeName);
-    Process process = nodeInfo.processMap.get(serverType);
-    log.debug("Destroying process with pid {} for {}", process.pid(), nodeInfo.ip);
-    killProcess(process.pid());
+    if (nodeInfo != null) {
+      Process process = nodeInfo.processMap.get(serverType);
+      if (process != null) {
+        log.debug("Destroying process with pid {} for {}", process.pid(), nodeInfo.ip);
+        killProcess(process.pid());
+      }
+    }
+  }
+
+  public void startProcess(
+      UUID universeUuid, String nodeName, UniverseTaskBase.ServerType serverType) {
+    Universe universe = Universe.getOrBadRequest(universeUuid);
+    NodeInfo nodeInfo = nodesByNameMap.get(nodeName);
+    UserIntent userIntent = universe.getCluster(nodeInfo.placementUUID).userIntent;
+    startProcessForNode(userIntent, serverType, nodeInfo);
+  }
+
+  public boolean isProcessRunning(String nodeName, UniverseTaskBase.ServerType serverType) {
+    NodeInfo nodeInfo = nodesByNameMap.get(nodeName);
+    if (nodeInfo == null) {
+      return false;
+    }
+    return nodeInfo.processMap.containsKey(serverType);
   }
 
   public void checkAllProcessesAlive() {
